pyproject.toml
[project]
name = "collector-of-collectors-v2ray"
version = "0.1.0"
description = "Collect, deduplicate, test, and serve V2Ray subscription configs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "httpx[socks]>=0.27.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
